only install package once

+++ launch_chffrplus.sh
@@ -77,7 +77,10 @@ function launch {
 
   # hardware specific init
   if [ -f /AGNOS ]; then
-    pip install soundfile openwakeword silero_vad groq openai librosa geopy
+    if [ ! -f /data/params/d/AceAutoPip ]; then
+      pip install soundfile openwakeword silero_vad groq openai librosa geopy
+      echo -n 1 > /data/params/d/AceAutoPip
+    fi
     agnos_init
   fi
